refactor: pull tower-related things into service.rs
